+++ extensions/threads/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Threads Changelog
 
+## [Update] - {PR_MERGE_DATE}
+
+- Updated the `Download Threads Media` command to download media from Threads.
+
 ## [Update] - 2025-04-28
 
 - Updated the `Download Threads Media` command to download media from Threads.